llama3/requirements.txt
fastapi
llama-cpp-python
uvicorn
python-multipart
sentence_transformers